refactor(models): Add change_source types to public API
Add 6 change_source types to vibe3.models public API and update 3 analysis
module imports to use package-level import path.

Changes:
- src/vibe3/models/__init__.py: Add imports and __all__ entries for
  BranchSource, ChangeSource, ChangeSourceType, CommitSource, PRSource,
  UncommittedSource
- src/vibe3/analysis/inspect_query_service.py: Update import to use
  vibe3.models instead of vibe3.models.change_source
- src/vibe3/analysis/serena_service.py: Update import to use
  vibe3.models instead of vibe3.models.change_source
- src/vibe3/analysis/change_scope_service.py: Update TYPE_CHECKING import
  to use vibe3.models instead of vibe3.models.change_source

This is a purely additive change with no breaking changes. Existing direct
submodule imports continue to work.

Ref: #1703
